LGTK

+++ useradmin-svc/src/main/java/gov/dhs/uscis/odos2/useradmin/repository/UserRolesRepository.java
@@ -2,8 +2,9 @@ package gov.dhs.uscis.odos2.useradmin.repository;
 
 import gov.dhs.uscis.odos2.useradmin.model.UserRoles;
 import org.springframework.data.repository.CrudRepository;
+import java.util.UUID;
 
-public interface UserRolesRepository extends CrudRepository<UserRoles, Long> {
+public interface UserRolesRepository extends CrudRepository<UserRoles, Integer> {
     UserRoles save(UserRoles userrole);
 
     void delete(UserRoles userrole);
